rename ratelimit to freshness, etc
